Move driver file into handler package
