+++ src/internal/handlers/main_handler.go
@@ -1,4 +1,4 @@
-package main
+package handlers
 
 import (
 	"context"
@@ -8,10 +8,9 @@ import (
 	"github.com/jonleeyz/bball8bot/internal/handlers/messages"
 )
 
-// handleUpdate parses the input Update object, and responds accordingly.
+// HandleUpdate parses the input Update object, and responds accordingly.
 // No error is returned from this function, any error is logged but is not fatal; other updates can still be processed.
-func handleUpdate(ctx context.Context, update *tgbotapi.Update, bot *tgbotapi.BotAPI) {
-	// TODO @jonlee: Update, placeholder, just to ensure that callback queries are answered.
+func HandleUpdate(ctx context.Context, update *tgbotapi.Update, bot *tgbotapi.BotAPI) {
 	if update.CallbackQuery != nil {
 		handler := callbacks.Init(bot)
 		handler.Handle(ctx, update)
